# docs/swagger.yaml
definitions:
  model.ExportFileListDTO:
    properties:
      filename:
        description: Type TAX_BILL, EXPORT_XML
        type: string
      filepath:
        type: string
      modifiedTime:
        description: 修改时间
        type: string
      size:
        description: 文件大小 bytes
        type: integer
    type: object
  model.FileResendRequest:
    properties:
      customPath:
        description: CustomPath 自定义路径(默认：不需要, 特殊情况，用户可能需要指当前需要重新触发监听的export xml 文件所在的路径)
        type: string
      files:
        description: Files 需要重新发送的文件名，多个文件名需要在同一路径中。监听路径或者备份路径
        items:
          type: string
        type: array
      inWatchPath:
        description: InWatchPath 是否是监听路径中（默认：false,即在备份路径中）
        type: boolean
    required:
    - files
    type: object
  model.ResponseError:
    properties:
      errors:
        description: Error messages
        items:
          type: string
        type: array
      status:
        description: Status success, fail
        type: string
    type: object
  model.SearchFileRequest:
    properties:
      declareCountry:
        description: DeclareCountry NL, BE
        type: string
      filenames:
        description: Filenames Support use Job number
        items:
          type: string
        type: array
      month:
        description: 'Month exp: 09'
        type: string
      type:
        description: Type TAX_BILL, EXPORT_XML
        type: string
      year:
        description: 'Year exp: 2022'
        type: string
    required:
    - declareCountry
    - filenames
    - type
    type: object
  model.SearchFileResult:
    properties:
      filename:
        type: string
      filepath:
        type: string
      searchText:
        type: string
      type:
        description: Type TAX_BILL, EXPORT_XML
        type: string
    type: object
info:
  contact: {}
paths:
  /download/pdf/{origin}/{target}:
    get:
      consumes:
      - application/json
      description: 通过申报国家确定税金文件路径
      parameters:
      - description: 下载的源文件名,不带文件后缀
        in: path
        name: origin
        required: true
        type: string
      - description: 下载文件后，将文件重命名的文件名，没有后缀将自动添加pdf作为后缀
        in: path
        name: target
        required: true
        type: string
      - description: 申报国家(BE|NL),默认为NL
        in: query
        name: dc
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 下载税金单文件
      tags:
      - download
  /download/xml/{dc}/{filename}:
    get:
      consumes:
      - application/json
      description: 通过文件名前缀确定文件路径
      parameters:
      - description: 申报国家(BE|NL)
        in: path
        name: dc
        required: true
        type: string
      - description: export文件的文件名
        in: path
        name: filename
        required: true
        type: string
      - description: 是否下载，1表示直接下载
        in: query
        name: download
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: 下载export文件
      tags:
      - download
  /export/list/{dc}:
    get:
      consumes:
      - application/json
      description: 通过指定的申报国家获取其当前Export监听路径下的文件列表
      parameters:
      - description: 申报国家(BE|NL)
        in: path
        name: dc
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExportFileListDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseError'
      summary: 获取Export监听路径下当前文件列表
      tags:
      - export
  /export/remover/{dc}:
    post:
      consumes:
      - application/json
      description: 页面选取Export文件并发送文件完整路径，将Export文件重新移入Export监听路径中，触发文件的CREATE监听，从而重新发送
      parameters:
      - description: 申报国家(BE|NL)
        in: path
        name: dc
        required: true
        type: string
      - description: 需要重新发送的文件完整路径
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.FileResendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ExportFileListDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseError'
      summary: 重新发送Export XML 文件
      tags:
      - export
  /search/file:
    post:
      consumes:
      - application/json
      description: 可检索税金单文件以及export报关结果文件，可使用文件名部分做模糊匹配。建议使用Job number 进行检索
      parameters:
      - description: 检索内容
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/model.SearchFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SearchFileResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseError'
      summary: 搜索文件
      tags:
      - search
swagger: "2.0"